chore(root): cleans up scripts
