+++ .github/workflows/publish.yml
@@ -137,7 +137,7 @@ jobs:
           fi
 
           if gh release view "$VERSION" > /dev/null 2>&1; then
-            sh scripts/upload-release-assets.sh "$VERSION" "$TARBALL" "$SIGSTORE_BUNDLE"
+            nub scripts/release/index.ts assets "$VERSION" "$TARBALL" "$SIGSTORE_BUNDLE"
             exit 0
           fi
