+++ codecov.yml
@@ -13,4 +13,8 @@ coverage:
     - "src/**/types.ts"
     - "src/constants.ts"
     - "src/**/constants.ts"
-    - "scripts/**"
+    - "scripts/**/types.ts"
+    - "scripts/**/constants.ts"
+    - "scripts/build/**"
+    - "scripts/run/**"
+    - "scripts/release/index.ts"
