--- scripts/ci/constants.js
@@ -1,9 +1,0 @@
-export const TOOL_OUTPUT_KEYS = {
-  nodeAlpineImage: "node_alpine_image",
-  nodeSlimImage: "node_slim_image",
-  nodeVersion: "node_version",
-  nubVersion: "nub_version",
-};
-
-export const RELEASE_VERSION_PATTERN =
-  /^v?\d+\.\d+\.\d+(?:-([0-9A-Za-z-]+)(?:\.[0-9A-Za-z-]+)*)?(?:\+[0-9A-Za-z.-]+)?$/;
