--- scripts/constants.ts
@@ -1,28 +1,0 @@
-export const BIN_OUTPUT_DIR = "artifacts";
-export const BIN_OUTPUT_FILE = `${BIN_OUTPUT_DIR}/codependence`;
-export const BIN_ENTRY_SOURCE_FILE = "scripts/bin/entry.ts";
-export const BIN_ENTRY_FILE = `${BIN_OUTPUT_DIR}/entry.ts`;
-export const BIN_RUNTIME_NAME = "codependence-runtime";
-export const BIN_RUNTIME_SOURCE_FILE = "src/cli/index.ts";
-export const BIN_RUNTIME_DIR = `${BIN_OUTPUT_DIR}/node_modules/${BIN_RUNTIME_NAME}`;
-export const BIN_RUNTIME_PACKAGE_FILE = `${BIN_RUNTIME_DIR}/package.json`;
-export const BIN_RUNTIME_TYPES_FILE = `${BIN_RUNTIME_DIR}/index.d.ts`;
-export const BIN_BUNDLE_FILE = `${BIN_RUNTIME_DIR}/index.js`;
-export const BIN_BUNDLE_ARGS = [
-  BIN_RUNTIME_SOURCE_FILE,
-  "--file",
-  BIN_BUNDLE_FILE,
-  "--format",
-  "esm",
-  "--platform",
-  "node",
-  "--minify",
-];
-export const BIN_BUILD_ARGS = [
-  "build",
-  BIN_ENTRY_FILE,
-  "-o",
-  BIN_OUTPUT_FILE,
-  "--dynamic",
-  "--no-keep-c",
-];
