+++ scripts/release/constants.ts
@@ -2,6 +2,8 @@ import type { PreRelease, ReleaseIncrement } from "./types";
 
 export const TAG_VERSION_PATTERN = /^\d+\.\d+\.\d+(-[0-9A-Za-z.-]+)?(\+[0-9A-Za-z.-]+)?$/;
 export const RELEASE_VERSION_PATTERN = /\d+\.\d+\.\d+(?:-[0-9A-Za-z.-]+)?(?:\+[0-9A-Za-z.-]+)?/g;
+export const PUBLISHED_RELEASE_VERSION_PATTERN =
+  /^v?\d+\.\d+\.\d+(?:-([0-9A-Za-z-]+)(?:\.[0-9A-Za-z-]+)*)?(?:\+[0-9A-Za-z.-]+)?$/;
 export const STABLE_VERSION_PATTERN = /^\d+\.\d+\.\d+$/;
 export const PRE_RELEASE_VERSION_PATTERN = /^\d+\.\d+\.\d+-[0-9A-Za-z.-]+(?:\+[0-9A-Za-z.-]+)?$/;
 export const PRE_RELEASES = new Set<PreRelease>(["alpha", "beta", "rc"]);
@@ -49,3 +51,14 @@ export const FORMULA_BODY = [
   "  end",
   "end",
 ];
+export const TEST_PUBLISHED_RELEASE_COMMANDS = new Set([
+  "build-release-image",
+  "compatibility-check",
+  "resolve-version",
+  "run-e2e",
+  "run-npm-smoke",
+  "summary",
+  "verify-installation",
+  "wait-for-npm",
+  "write-report",
+]);
