+++ tests/integration/release-workflows.test.ts
@@ -69,11 +69,8 @@ describe("release workflows", () => {
   test("uses current release tooling when retrying old tags", () => {
     const homebrew = readWorkflow("homebrew.yml");
     assert.ok(homebrew.includes('ref: "${{ github.workflow_sha }}"'));
-    assert.ok(homebrew.includes("scripts/upload-release-assets.sh"));
-    assert.ok(homebrew.includes("scripts/release/index.ts"));
-    assert.ok(homebrew.includes("scripts/release/constants.ts"));
-    assert.ok(homebrew.includes("scripts/release/types.ts"));
-    assert.ok(homebrew.includes("scripts/release/utils.ts"));
+    assert.ok(homebrew.includes("scripts/release/*"));
+    assert.ok(homebrew.includes("scripts/ci/tool-versions.js"));
     assert.ok(homebrew.includes("src/observability/*"));
     assert.ok(homebrew.includes("src/dx/constants.ts"));
     assert.ok(homebrew.includes("src/dx/output/*"));
@@ -102,13 +99,14 @@ describe("release workflows", () => {
   test("tests the exact published release", () => {
     const releaseTest = readWorkflow("test-release.yml");
     assert.ok(releaseTest.includes("inputs.version || github.event.release.tag_name"));
+    assert.ok(releaseTest.includes("nub scripts/release/index.ts test-published"));
   });
 
   test("publishes release assets immutably", () => {
     const files = [
       readWorkflow("homebrew.yml"),
       readWorkflow("publish.yml"),
-      readScript("upload-release-assets.sh"),
+      readScript("release/utils.ts"),
     ];
     files.forEach((file) => assert.ok(!file.includes("--clobber")));
     assert.ok(files.at(-1).includes("Release asset digest mismatch"));
@@ -132,7 +130,6 @@ describe("release workflows", () => {
     const homebrew = readWorkflow("homebrew.yml");
     const brewScript = readScript("release/utils.ts");
     assert.ok(!homebrew.includes("gh pr merge --auto"));
-    assert.ok(!brewScript.includes("allow_auto_merge"));
     assert.ok(brewScript.includes("Automated formula update."));
   });
 });
