+++ tests/unit/scripts/upload-release-assets.test.ts
@@ -1,38 +1,31 @@
 import { afterEach, describe, test } from "node:test";
 import assert from "node:assert/strict";
 import { createHash } from "node:crypto";
-import { spawnSync } from "node:child_process";
-import { chmodSync, mkdirSync, mkdtempSync, readFileSync, rmSync, writeFileSync } from "node:fs";
-import { tmpdir } from "node:os";
+import { mkdirSync, mkdtempSync, rmSync, writeFileSync } from "node:fs";
 import { join } from "node:path";
-import { fileURLToPath } from "node:url";
+import { runUploadReleaseAssetsCli } from "../../../scripts/release";
+import type { GitResult, ReleaseRunner } from "../../../scripts/release";
 
-const SCRIPT_PATH = fileURLToPath(
-  new URL("../../../scripts/upload-release-assets.sh", import.meta.url),
-);
-const FAKE_GH = `#!/bin/sh
-printf '%s\n' "$*" >> "$FAKE_GH_LOG"
-if [ "$1" = "api" ] && [ "$2" = "--paginate" ]; then
-  printf '%s' "$FAKE_RELEASE_JSON"
-  exit 0
-fi
-if [ "$1" = "api" ] && [ "$2" = "--header" ]; then
-  printf '%s' "$FAKE_ASSET_BODY"
-  exit 0
-fi
-`;
 const RELEASE_ID = 123;
 const UPLOAD_URL =
   "https://uploads.github.com/repos/yowainwright/codependence/releases/123/assets{?name,label}";
+const TEMP_ROOT = join(import.meta.dirname, ".tmp-release-assets");
 const tempDirectories = new Set<string>();
 
-type Fixture = { assetPath: string; env: NodeJS.ProcessEnv; logPath: string };
+type Fixture = {
+  assetPath: string;
+  assetBody: string;
+  env: NodeJS.ProcessEnv;
+  log: string[];
+  releaseJson: string;
+};
 type FixtureOptions = {
   assetContent?: string;
   assetName?: string;
   existingAssetBody?: string;
   publishedDigest?: string | null;
   releaseExists?: boolean;
+  uploadUrl?: string;
 };
 
 function createReleaseJson(
@@ -54,18 +47,13 @@ function createReleaseJson(
 }
 
 function createEnvironment(
-  binPath: string,
-  logPath: string,
   releaseJson: string,
-  assetBody = "",
+  assetBody: string,
 ) {
-  const path = `${binPath}:${process.env.PATH ?? ""}`;
   return Object.assign({}, process.env, {
     FAKE_ASSET_BODY: assetBody,
-    FAKE_GH_LOG: logPath,
     FAKE_RELEASE_JSON: releaseJson,
     GITHUB_REPOSITORY: "yowainwright/codependence",
-    PATH: path,
   });
 }
 
@@ -84,65 +72,97 @@ function createFixtureWithOptions({
   existingAssetBody = "",
   publishedDigest,
   releaseExists = true,
+  uploadUrl = UPLOAD_URL,
 }: FixtureOptions = {}): Fixture {
-  const root = mkdtempSync(join(tmpdir(), "codependence-release-assets-"));
-  const binPath = join(root, "bin");
+  mkdirSync(TEMP_ROOT, { recursive: true });
+  const root = mkdtempSync(join(TEMP_ROOT, "fixture-"));
   const assetPath = join(root, assetName);
-  const logPath = join(root, "gh.log");
   const assets =
     publishedDigest === undefined ? [] : [{ name: assetName, digest: publishedDigest }];
   tempDirectories.add(root);
-  mkdirSync(binPath);
   writeFileSync(assetPath, assetContent);
-  writeFileSync(logPath, "");
-  writeFileSync(join(binPath, "gh"), FAKE_GH);
-  chmodSync(join(binPath, "gh"), 0o755);
-  const releaseJson = createReleaseJson(assets, releaseExists);
-  const env = createEnvironment(binPath, logPath, releaseJson, existingAssetBody);
-  return { assetPath, env, logPath };
+  const releaseJson = createReleaseJson(assets, releaseExists).replace(UPLOAD_URL, uploadUrl);
+  const env = createEnvironment(releaseJson, existingAssetBody);
+  return { assetPath, assetBody: existingAssetBody, env, log: [], releaseJson };
 }
 
 function createFixture(publishedDigest?: string | null, releaseExists = true): Fixture {
   return createFixtureWithOptions({ publishedDigest, releaseExists });
 }
 
-const runUpload = ({ assetPath, env }: Fixture) =>
-  spawnSync("sh", [SCRIPT_PATH, "v1.2.3", assetPath], { encoding: "utf8", env });
+function runUpload(fixture: Fixture) {
+  let stdout: string[] = [];
+  const runner = createRunner(fixture);
+  const logger = {
+    log(value: unknown) {
+      stdout = stdout.concat(String(value));
+    },
+  };
+
+  try {
+    const status = runUploadReleaseAssetsCli({
+      argv: ["v1.2.3", fixture.assetPath],
+      env: fixture.env,
+      logger,
+      runner,
+    });
+    return { status, stderr: "", stdout: `${stdout.join("\n")}\n` };
+  } catch (error) {
+    const message = error instanceof Error ? error.message : String(error);
+    return { status: 1, stderr: message, stdout: stdout.join("\n") };
+  }
+}
+
+function createRunner(fixture: Fixture): ReleaseRunner {
+  return (command, args) => {
+    const line = [command].concat(Array.from(args)).join(" ");
+    fixture.log = fixture.log.concat(line);
+    return runnerResult(fixture, args);
+  };
+}
+
+function runnerResult(fixture: Fixture, args: readonly string[]): GitResult {
+  const readsReleases = args.includes("--paginate");
+  if (readsReleases) return { status: 0, stdout: fixture.releaseJson, stderr: "" };
+
+  const readsAsset = args.includes("Accept: application/octet-stream");
+  if (readsAsset) return { status: 0, stdout: fixture.assetBody, stderr: "" };
+
+  return { status: 0, stdout: "", stderr: "" };
+}
 
 afterEach(() => {
   tempDirectories.forEach((directory) => rmSync(directory, { recursive: true }));
   tempDirectories.clear();
+  rmSync(TEMP_ROOT, { recursive: true, force: true });
 });
 
-describe("scripts/upload-release-assets", () => {
+describe("scripts/release assets", () => {
   test("uploads a missing asset", () => {
     const fixture = createFixture();
     const result = runUpload(fixture);
-    const log = readFileSync(fixture.logPath, "utf8");
     assert.strictEqual((result.status), 0);
     assert.strictEqual((result.stdout), "123\n");
-    assert.ok((log).includes("api --paginate --slurp"));
-    assert.ok((log).includes("api --method POST"));
-    assert.ok((log).includes("releases/123/assets?name=codependence.tgz"));
+    assert.ok(fixture.log.some((line) => line.includes("api --paginate --slurp")));
+    assert.ok(fixture.log.some((line) => line.includes("api --method POST")));
+    assert.ok(fixture.log.some((line) => line.includes("releases/123/assets?name=codependence.tgz")));
   });
 
   test("skips an existing asset with the expected digest", () => {
     const digest = createHash("sha256").update("release asset").digest("hex");
     const fixture = createFixture(`sha256:${digest}`);
     const result = runUpload(fixture);
-    const log = readFileSync(fixture.logPath, "utf8");
     assert.strictEqual((result.status), 0);
     assert.strictEqual((result.stdout), "123\n");
-    assert.ok(!(log).includes("api --method POST"));
+    assert.ok(!fixture.log.some((line) => line.includes("api --method POST")));
   });
 
   test("rejects an existing asset with a different digest", () => {
     const fixture = createFixture("sha256:unexpected");
     const result = runUpload(fixture);
-    const log = readFileSync(fixture.logPath, "utf8");
     assert.strictEqual((result.status), 1);
     assert.ok((result.stderr).includes("Release asset digest mismatch: codependence.tgz"));
-    assert.ok(!(log).includes("api --method POST"));
+    assert.ok(!fixture.log.some((line) => line.includes("api --method POST")));
   });
 
   test("skips an existing Sigstore bundle with the same subject digest", () => {
@@ -154,11 +174,10 @@ describe("scripts/upload-release-assets", () => {
       publishedDigest: "sha256:unexpected",
     });
     const result = runUpload(fixture);
-    const log = readFileSync(fixture.logPath, "utf8");
     assert.strictEqual((result.status), 0);
     assert.strictEqual((result.stdout), "123\n");
-    assert.ok((log).includes("api --header Accept: application/octet-stream"));
-    assert.ok(!(log).includes("api --method POST"));
+    assert.ok(fixture.log.some((line) => line.includes("api --header Accept: application/octet-stream")));
+    assert.ok(!fixture.log.some((line) => line.includes("api --method POST")));
   });
 
   test("rejects an existing Sigstore bundle with a different subject digest", () => {
@@ -187,5 +206,22 @@ describe("scripts/upload-release-assets", () => {
     const result = runUpload(fixture);
     assert.strictEqual((result.status), 1);
     assert.ok((result.stderr).includes("Release not found: v1.2.3"));
+  });
+
+  test("rejects an unexpected release upload URL", () => {
+    const fixture = createFixtureWithOptions({
+      uploadUrl: "https://uploads.github.com/repos/other/project/releases/123/assets{?name,label}",
+    });
+    const result = runUpload(fixture);
+    assert.strictEqual((result.status), 1);
+    assert.ok((result.stderr).includes("Unexpected release upload URL"));
+  });
+
+  test("rejects missing CLI arguments", () => {
+    assert.throws(() => runUploadReleaseAssetsCli({ argv: [], env: {} }), /Release tag is required/);
+    assert.throws(
+      () => runUploadReleaseAssetsCli({ argv: ["v1.2.3"], env: {} }),
+      /At least one release asset is required/,
+    );
   });
 });
